Implemented light handler
The bulb turns on automatically if there is not enough light being
picked up by the light sensor. The bulb can also be activated by
pressing a button. This functionality can be switched between by
pressing the button for a few seconds.

The loop function has also been updated to be delayed by milliseconds,
but doesn't include the lightControl function as it prevents the button
from working consistently during delays.